ElaWidgetTools Commit 2488f5d4 까지 적용

+++ ElaWidgetTools/DeveloperComponents/ElaCheckBoxStyle.cpp
@@ -85,7 +85,7 @@ void ElaCheckBoxStyle::drawControl(ControlElement element, const QStyleOption* o
                 QFont iconFont = QFont("ElaAwesome");
                 iconFont.setPixelSize(_pCheckIndicatorWidth * 0.75);
                 painter->setFont(iconFont);
-                painter->drawText(checkRect, Qt::AlignCenter, QChar((unsigned short)ElaIconType::Check));
+                painter->drawText(checkRect, Qt::AlignCenter, QChar((int)ElaIconType::Check));
                 painter->restore();
             }
             else if (bopt->state.testFlag(QStyle::State_NoChange))
